Final Submission
While the program does not run correctly, this is due to a lapse in development, where I failed to create a correct timeline of events that will give the outputs required. The specific issue is that I was incorrectly parsing information through the data structure, requiring a complete rework of the entire Queue class, which in turn would affect every other classes structure.

Also added a .gitignore to remove Word Docs, pdfs, zip files and class files.